Add plotDataHttpProvider service
This allows you to override the usual use of $http to get plot data
from a server in <plot-data> directives with a SRC attribute.  Instead
of using $http, you can pass in any service that supports a get()
method that takes a URL and handles responses in the usual
success/error promise fashion.  This is useful for client-side
caching.